+++ src/utils/quadrique_operation.c
@@ -6,27 +6,27 @@
 /*   By: maandria <maandria@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 15:54:25 by maandria          #+#    #+#             */
-/*   Updated: 2025/03/03 17:17:12 by maandria         ###   ########.fr       */
+/*   Updated: 2025/03/18 14:25:31 by maandria         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/minirt.h"
 
 /*********** calcul du discriminant Delta **********************/
-float	op_delta(float A, float B, float C)
+double	op_delta(double A, double B, double C)
 {
-	float	delta;
+	double	delta;
 
 	delta = (B * B) - (4 * (A * C));
 	return (delta);
 }
 
 /********** calcul des solutions possible de t *************/
-float	op_solution_quad(float A, float B, float C)
+double	op_solution_quad(double A, double B, double C)
 {
-	float	result;
-	float	delta;
-	float	tmp;
+	double	result;
+	double	delta;
+	double	tmp;
 
 	delta = op_delta(A, B, C);
 	result = -1;
